Simplified application.js to speed up loading of Stub page for chrome endorsement extension Candidate suggestion editor Cleared 55 problems (6 errors, 49 warnings) in unrelated classes Renamed index.jsx to ItemActionBar.jsx
